move python code to target dir
