+++ python/tvm/topi/x86/conv2d_int8.py
@@ -30,7 +30,7 @@ from ..nn.depthwise_conv2d import _get_workload as _get_depthwise_conv2d_workloa
 from ..utils import get_const_tuple, traverse_inline
 from .. import nn
 from . import conv2d_avx_1x1, conv2d_avx_common
-from .utils import target_has_sse42
+from tvm.target.x86 import target_has_sse42
 
 
 def _get_default_config_int8(
