+++ src/meta_schedule/space_generator/space_generator.cc
@@ -24,14 +24,14 @@ namespace meta_schedule {
 String GetRuleKindFromTarget(const Target& target) {
   if (target->kind->name == "llvm") {
     static const PackedFunc* f_check_vnni =
-        runtime::Registry::Get("tvm.topi.x86.utils.target_has_vnni");
+        runtime::Registry::Get("tvm.target.x86.target_has_vnni");
     ICHECK(f_check_vnni != nullptr) << "The `target_has_vnni` func is not in tvm registry.";
     if (target->GetAttr<String>("mcpu") &&
         (*f_check_vnni)(target->GetAttr<String>("mcpu").value())) {
       return "vnni";
     } else {
       static const PackedFunc* f_check_avx512 =
-          runtime::Registry::Get("tvm.topi.x86.utils.target_has_avx512");
+          runtime::Registry::Get("tvm.target.x86.target_has_avx512");
       ICHECK(f_check_avx512 != nullptr) << "The `target_has_avx512` func is not in tvm registry.";
       if (target->GetAttr<String>("mcpu") &&
           (*f_check_avx512)(target->GetAttr<String>("mcpu").value())) {
